# docs/reference/api/v1/pipeline-service.openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RunStartRequest": {
        "description": "Input to POST /api/v1/pipeline/runs \u2014 run a sequence of stages.",
        "properties": {
          "base_run_fallback_stages": {
            "description": "Stages upstream lidos do run base (ADR-291/ADR-303)",
            "items": {
              "type": "string"
            },
            "title": "Base Run Fallback Stages",
            "type": "array"
          },
          "base_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run base para fallback run-pinado em from_stage (ADR-291/ADR-303)",
            "title": "Base Run Id"
          },
          "config_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Dir"
          },
          "incremental": {
            "default": false,
            "title": "Incremental",
            "type": "boolean"
          },
          "incremental_doc_paths": {
            "items": {
              "type": "string"
            },
            "title": "Incremental Doc Paths",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "skip_llm": {
            "default": true,
            "title": "Skip Llm",
            "type": "boolean"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Stages",
            "type": "array"
          },
          "stop_on_error": {
            "default": true,
            "title": "Stop On Error",
            "type": "boolean"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          },
          "workspace_root": {
            "title": "Workspace Root",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "workspace_id",
          "workspace_root",
          "stages"
        ],
        "title": "RunStartRequest",
        "type": "object"
      },
      "RunSummaryResponse": {
        "description": "Aggregated result of a multi-stage run.",
        "properties": {
          "failed_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Stage"
          },
          "finished_at": {
            "title": "Finished At",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageExecuteResponse"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "workspace_id",
          "success",
          "started_at",
          "finished_at",
          "stages"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
      "ServiceHealthResponse": {
        "properties": {
          "service": {
            "default": "pipeline-service",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "title": "ServiceHealthResponse",
        "type": "object"
      },
      "StageExecuteRequest": {
        "description": "Input to POST /api/v1/pipeline/stages/{stage}/execute.\n\nBackend owns the workspace root on disk and passes it by path \u2014 the\npipeline-service runs the stage in-process using `pipeline.orchestrator`.\nArtifacts are written to/read from that root by the stage runners.",
        "properties": {
          "base_run_fallback_stages": {
            "description": "Stages upstream lidos do run base (ADR-291/ADR-303)",
            "items": {
              "type": "string"
            },
            "title": "Base Run Fallback Stages",
            "type": "array"
          },
          "base_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run base para fallback run-pinado em from_stage (ADR-291/ADR-303)",
            "title": "Base Run Id"
          },
          "config_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Dir"
          },
          "incremental": {
            "default": false,
            "title": "Incremental",
            "type": "boolean"
          },
          "incremental_doc_paths": {
            "items": {
              "type": "string"
            },
            "title": "Incremental Doc Paths",
            "type": "array"
          },
          "run_id": {
            "description": "Backend-issued PipelineRun UUID",
            "title": "Run Id",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          },
          "workspace_root": {
            "description": "Absolute path to tenant root",
            "title": "Workspace Root",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "workspace_id",
          "workspace_root"
        ],
        "title": "StageExecuteRequest",
        "type": "object"
      },
      "StageExecuteResponse": {
        "description": "Result of a single stage execution.",
        "properties": {
          "attempts": {
            "default": 1,
            "title": "Attempts",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "duration_ms": {
            "default": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "stage",
          "success"
        ],
        "title": "StageExecuteResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP boundary for pipeline execution (A6f.1 \u00b7 ADR-112).",
    "title": "Mathoms Pipeline Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/pipeline/runs": {
      "post": {
        "description": "Execute the requested stage sequence synchronously.\n\nBackend caller is responsible for run_id generation and DB persistence;\nthis service is stateless and returns the aggregated result only.\nBoundary externo: aceita nomes legacy ou descritivos (ADR-093).",
        "operationId": "start_run_api_v1_pipeline_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/pipeline/stages/{stage}/execute": {
      "post": {
        "description": "Execute a single pipeline stage and return the result.\n\nBoundary externo: aceita nome legacy (\"E3\") ou descritivo (ADR-093);\nexecuta e responde sempre com o descritivo can\u00f4nico.",
        "operationId": "execute_stage_api_v1_pipeline_stages__stage__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "stage",
            "required": true,
            "schema": {
              "title": "Stage",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Stage",
        "tags": [
          "stages"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceHealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}